Relocate schema instrumentation to `utils/schemaInstrumentation.ts`.
The `requestPipelineAPI.ts`'s purpose was originally to keep typings by
themselves.  It was compiled using a separate TypeScript compilation
stage to avoid some circular dependencies within the repository itself.

However, it still proved to be problematic since it required external
packages which depended on the entire `apollo-server-core` just to utilize
those types (e.g. plugins!)

The work in #2990 offloaded the types to their own package that
could be depended on but the assertion in [[1]] correctly notes that
introducing new functionality, which is largely incompatible with the
original intent of the `requestPipelineAPI` file (even though it is now
deprecated) is largely a step backward.

Therefore, this moves the functionality to a new file called
`schemaInstrumentation`, as suggested in the following comment.

[1]: https://github.com/apollographql/apollo-server/pull/3988/files#r414666538
